move click logic to left sider out comp instead of in parent
