+++ src/components/LeftSidebar/index.tsx
@@ -1,5 +1,6 @@
 import NewProjectButton from 'components/NewProjectButton';
-import React from 'react';
+import { useProject } from 'providers/Project/projectHooks';
+import React, { useEffect, useRef } from 'react';
 import { SXStyles } from 'src/types';
 import { Box } from 'theme-ui';
 import ProjectsList from './ProjectsList';
@@ -12,9 +13,24 @@ const styles: SXStyles = {
   },
 };
 
-const LeftSidebar = ({ reference }: { reference: any }) => {
+const LeftSidebar = () => {
+  const { toggleProjectsSidebar } = useProject();
+  const ref = useRef(null);
+
+  useEffect(() => {
+    const handleClickOutside = (event: Event) => {
+      if (ref.current && !ref.current.contains(event.target)) {
+        toggleProjectsSidebar();
+      }
+    };
+    document.addEventListener('click', handleClickOutside, true);
+    return () => {
+      document.removeEventListener('click', handleClickOutside, true);
+    };
+  }, []);
+
   return (
-    <Box sx={styles.root} ref={reference}>
+    <Box sx={styles.root} ref={ref}>
       <Box mb={12}>
         <NewProjectButton
           label="Create New Project"
